+++ src/main/java/co/invest72/investment/console/output/InvestmentResultPrinter.java
@@ -1,17 +1,18 @@
 package co.invest72.investment.console.output;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import co.invest72.investment.presentation.response.MonthlyInvestmentResult;
 
 public interface InvestmentResultPrinter {
-	void printTotalPrincipal(int amount);
+	void printTotalPrincipal(BigDecimal amount);
 
-	void printInterest(int amount);
+	void printInterest(BigDecimal amount);
 
-	void printTax(int amount);
+	void printTax(BigDecimal amount);
 
-	void printTotalProfit(int amount);
+	void printTotalProfit(BigDecimal amount);
 
 	void printMonthlyInvestmentResults(List<MonthlyInvestmentResult> monthlyInvestmentResults);
 }
